requirements.txt
# Python 3.10+ required
# RAG Chatbot: FastAPI + ChromaDB + Qwen (DashScope OpenAI-compatible)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.5.0
